**AncientTable**
Inputing relics will increase your **knowledge** eventually this will shine meaning it need an AncientRelic to advance to the next **Milestone**

**AncientTableMenu**

ConsumingSlot - gets ***Papyrus?***

RelicSlot - Gets an ***AncientRelic***

AwardSlot

AmuletSlot - To get an AncientMap

**Knowledge:**

When you use the ***AncientTable*** with a ***Papyrus***, you get some knowledge. The more knowledge you get a ***milestone*** that can give you a ***AncientMap***

*To know how much knowledge was give depends on the quality of item given back*

**KnowledgeMap**

This keeps track of the knowledge and when a ***Milestone*** is achieved, it activates something on the world

**Milestone**

Generates a structure in the world that was not there before, (to be done but from all sand_blocks into whats suppose to look like) inputing an **AncientRelic** will give you an ***AncientMap,*** the ***AncientTable***, will shine

**MilestoneTutankamon**

gets the AncientMap of his tomb, here there are mobs to defend and many traps, there is a warning of the ***TutankamonCurse***

**TutankamonCurse,** recieve damage when being in the overworld

2 options, not follow and go for the tech, straight to **MilestoneRa**

follow and go for the adventure accept the **TutankamonCurse**, get a massive reward but go into the **MilestoneAnubis**

**MilestoneAnubis**

Nether, we have to make it livable, so ancient_dark_sandstone, soulsand?

More research and discovery….

**AncientDebug**

In the meanwhile so we can see how much progress every AncientRelic is giving us and how much progress we have to the next milestone

**AncientRelic**

**Tutankamon_Gem**

**Anubis_Gem**

**Ra_Gems**

**Papyrus**

**AncientHieroglyph**

***AncientMap***

next standart, knowledge is done